split avatar movement into two systems
