+++ packages/engine/src/avatar/functions/spawnAvatarReceptor.ts
@@ -169,6 +169,7 @@ export const createAvatarController = (entity: Entity) => {
     addComponent(entity, AvatarControllerComponent, {
       cameraEntity: Engine.instance.currentWorld.cameraEntity,
       bodyCollider: undefined!,
+      desiredMovement: new Vector3(),
       movementEnabled: true,
       isJumping: false,
       isWalking: false,
